[wip] Acquire Workers through RuntimeManager.
FIXME: Getting Aborted while queue runner is waiting for next item

Implement the START_SCOPE and EXIT_SCOPE calls to trigger Worker
submission and to leave the scope of those Workers.

Clarify the difference between EXIT_SCOPE and STOP.

Note that Raptor does not actually provide an API for stopping Workers
from the Raptor master, so the CPI semantics do not map precisely to the
runtime logic.

- [X] Acquire the Raptor master task through the RuntimeManager.
- [X] Manage a CPI command queue translating CPI calls to
Raptor-backed
  Futures (RPTasks or RPC calls)
- [X] Make sure CPI Session is properly shut down. (Partially deferred)
- [X] Acquire the Worker(s) through CPI call to the RuntimeManager.
- [ ] Implement EXIT_SCOPE in the ScalemsMaster.
- [ ] Normalize the CPI usage to shut down everything cleanly and
  expeditiously. (In too many cases right now, tests take an improperly
  long time because of various timeouts.)
- [ ] Isolate CPIExecutor concurrent.futures.Executor support from
  asyncio support (avoid blocking the event loop by avoiding event loop
  usage in the main implementation)
- [ ] Separate `messages` for intercomponent communication from `cpi`
  calls and responses.

Ref #345, #377